feat(travelPost): 여행 게시글 이미지 S3 업로드/삭제 로직 적용 + 폴더 이름 변경(travel -> travelPost)
