+++ src/main/java/com/goteego/review/service/ReviewService.java
@@ -7,10 +7,10 @@ import com.goteego.review.dto.review.ReviewRequestDto;
 import com.goteego.review.dto.review.ReviewTargetDto;
 import com.goteego.review.dto.review.UserReviewSummaryDto;
 import com.goteego.review.repository.ReviewRepository;
-import com.goteego.travel.domain.ParticipationApplication;
-import com.goteego.travel.domain.TravelPost;
-import com.goteego.travel.repository.ParticipationApplicationRepository;
-import com.goteego.travel.repository.TravelPostRepository;
+import com.goteego.travelPost.domain.ParticipationApplication;
+import com.goteego.travelPost.domain.TravelPost;
+import com.goteego.travelPost.repository.ParticipationApplicationRepository;
+import com.goteego.travelPost.repository.TravelPostRepository;
 import com.goteego.user.domain.User;
 import com.goteego.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
