--- src/main/java/com/goteego/travel/domain/enumerate/PostType.java
@@ -1,5 +1,0 @@
-package com.goteego.travel.domain.enumerate;
-
-public enum PostType {
-    BEFORE, NOW
-}
