+++ src/main/java/com/goteego/travelPost/controller/ScheduleController.java
@@ -1,11 +1,10 @@
-package com.goteego.travel.controller;
+package com.goteego.travelPost.controller;
 
-import com.goteego.travel.domain.ParticipationApplication;
-import com.goteego.travel.domain.enumerate.ParticipationStatus;
-import com.goteego.travel.dto.participation.ParticipantStatusRequest;
-import com.goteego.travel.dto.participation.ParticipationApplicationResponseDto;
-import com.goteego.travel.dto.travel.BeforeTravelPostResponseDto;
-import com.goteego.travel.service.ScheduleService;
+import com.goteego.travelPost.domain.enumerate.ParticipationStatus;
+import com.goteego.travelPost.dto.participation.ParticipantStatusRequest;
+import com.goteego.travelPost.dto.participation.ParticipationApplicationResponseDto;
+import com.goteego.travelPost.dto.travel.BeforeTravelPostResponseDto;
+import com.goteego.travelPost.service.ScheduleService;
 import com.goteego.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,9 +20,9 @@ import java.util.List;
 @RequestMapping("/api/schedule")
 @RequiredArgsConstructor
 public class ScheduleController {
-    
+
     private final ScheduleService scheduleService;
-    
+
     /**
      * 내 일정 조회 (BEFORE 타입만)
      */
@@ -35,40 +34,40 @@ public class ScheduleController {
 
         Long currentUserId = user.getId();
         List<BeforeTravelPostResponseDto> schedules = scheduleService.getMySchedules(currentUserId, page, size);
-        
-        log.info("내 일정 조회 - userId: {}, page: {}, size: {}, totalCount: {}", 
+
+        log.info("내 일정 조회 - userId: {}, page: {}, size: {}, totalCount: {}",
                 currentUserId, page, size, schedules.size());
-        
+
         return ResponseEntity.ok(schedules);
     }
 
 
-
     //=================================================준형==========================================================//
+
     /**
      * 참가자 상태 변경 (승인/거절)
      * 여행 게시글 작성자가 참가 신청자의 상태를 승인하거나 거절하는 API
      */
     @PutMapping("/{travelPostId}/participants/{userId}")
     public ResponseEntity<ParticipationApplicationResponseDto> updateParticipantStatus(@PathVariable("travelPostId") Long travelPostId,
-                                                                            @PathVariable("userId") Long userId,
-                                                                            @RequestBody ParticipantStatusRequest requestDto,
-                                                                            @AuthenticationPrincipal User user) {
+                                                                                       @PathVariable("userId") Long userId,
+                                                                                       @RequestBody ParticipantStatusRequest requestDto,
+                                                                                       @AuthenticationPrincipal User user) {
 
         Long currentUserId = user.getId();
 
         ParticipationApplicationResponseDto participationApplicationResponseDto = scheduleService.updateParticipantStatus(travelPostId, userId,
                 ParticipationStatus.valueOf(requestDto.getStatus().toUpperCase()), currentUserId);
 
-        log.info("참가자 상태 변경 - travelPostId: {}, participantUserId: {}, status: {}, currentUserId: {}", 
+        log.info("참가자 상태 변경 - travelPostId: {}, participantUserId: {}, status: {}, currentUserId: {}",
                 travelPostId, userId, participationApplicationResponseDto.getStatus(), currentUserId);
-        
+
         return ResponseEntity.ok(participationApplicationResponseDto);
     }
-    
+
     /**
      * 특정 게시글의 참여자 목록 조회
-     * 
+     *
      * @param travelPostId 여행 게시글 ID
      * @return 참여자 정보 목록
      */
@@ -77,35 +76,25 @@ public class ScheduleController {
 
         List<ParticipationApplicationResponseDto> participants = scheduleService.getParticipants(travelPostId);
         log.info("참여자 목록 조회 - travelPostId: {}, participantCount: {}", travelPostId, participants.size());
-        
+
         return ResponseEntity.ok(participants);
     }
-    
+
     /**
      * 특정 게시글의 참가 신청 목록 조회
      */
     @GetMapping("/{travelPostId}/applications")
     public ResponseEntity<List<ParticipationApplicationResponseDto>> getParticipationApplications(
             @PathVariable("travelPostId") Long travelPostId) {
-        
+
         List<ParticipationApplicationResponseDto> applications = scheduleService.getParticipationApplications(travelPostId);
         log.info("참가 신청 목록 조회 - travelPostId: {}, applicationCount: {}", travelPostId, applications.size());
-        
+
         return ResponseEntity.ok(applications);
     }
 
 
-
     //===========================================================================================================//
-
-
-
-
-
-
-
-
-
 
 
     //=================================================재신==========================================================//
@@ -119,19 +108,19 @@ public class ScheduleController {
         log.info("승인된 참가자 수 조회 - travelPostId: {}, count: {}", travelPostId, count);
         return ResponseEntity.ok(count);
     }
-    
+
     /**
      * 특정 게시글의 대기 중인 참가자 수 조회
-     * 
+     *
      * @param travelPostId 여행 게시글 ID
      * @return 대기 중인 참가자 수
      */
     @GetMapping("/{travelPostId}/participants/pending/count")
     public ResponseEntity<Long> getPendingParticipantCount(@PathVariable("travelPostId") Long travelPostId) {
         Long count = scheduleService.getPendingParticipantCount(travelPostId);
-        
+
         log.info("대기 중인 참가자 수 조회 - travelPostId: {}, count: {}", travelPostId, count);
-        
+
         return ResponseEntity.ok(count);
     }
