+++ src/main/java/com/goteego/travelPost/domain/enumerate/PostType.java
@@ -1,0 +1,5 @@
+package com.goteego.travelPost.domain.enumerate;
+
+public enum PostType {
+    BEFORE, NOW
+}
