+++ src/main/java/com/goteego/travelPost/dto/participation/ParticipantStatusRequest.java
@@ -1,4 +1,4 @@
-package com.goteego.travel.dto.participation;
+package com.goteego.travelPost.dto.participation;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -10,6 +10,11 @@ import lombok.NoArgsConstructor;
 public class ParticipantStatusRequest {
     private String status;
 
-    public String getStatus() { return status; }
-    public void setStatus(String status) { this.status = status; }
+    public String getStatus() {
+        return status;
+    }
+
+    public void setStatus(String status) {
+        this.status = status;
+    }
 }
