+++ src/main/java/com/goteego/travelPost/dto/travel/BeforeTravelPostResponseDto.java
@@ -1,36 +1,50 @@
-package com.goteego.travel.dto.travel;
+package com.goteego.travelPost.dto.travel;
 
-import com.goteego.travel.domain.TravelPost;
+import com.goteego.travelPost.domain.TravelPost;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * 현지동행모집글 응답 DTO (NOW)
+ * 사전동행모집글 응답 DTO (BEFORE)
  * API 응답 시 사용되는 데이터 전송 객체
  */
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class NowTravelPostResponseDto {
+public class BeforeTravelPostResponseDto {
 
     private Long travelPostId;
     private String title;
+    private String content;
     private String location;
+    private Long viewCount;
+    private String startTime;
+    private String endTime;
     private AuthorDto author;
+    private Integer participants; // 현재까지 신청받은 인원 수
+    private Integer maxParticipants; // 모집 인원 제한
+    private String imageUrl;
     private String createdAt;
 
     /**
      * TravelPost 엔티티를 DTO로 변환
      */
-    public static NowTravelPostResponseDto from(TravelPost travelPost, String nickname) {
-        return NowTravelPostResponseDto.builder()
+    public static BeforeTravelPostResponseDto from(TravelPost travelPost, String nickname, Integer approvedParticipantCount, Long viewCount) {
+        return BeforeTravelPostResponseDto.builder()
                 .travelPostId(travelPost.getId())
                 .title(travelPost.getTitle())
+                .content(travelPost.getContent())
                 .location(travelPost.getLocation().name())
+                .viewCount(viewCount)
+                .startTime(travelPost.getStartTime().toString())
+                .endTime(travelPost.getEndTime().toString())
                 .author(createAuthorDto(travelPost.getUser(), nickname))
+                .participants(approvedParticipantCount != null ? approvedParticipantCount : 0)
+                .maxParticipants(travelPost.getRecruitLimit())
+                .imageUrl(travelPost.getImageUrl())
                 .createdAt(travelPost.getCreatedAt().toString())
                 .build();
     }
@@ -43,9 +57,6 @@ public class NowTravelPostResponseDto {
                 .userId(user.getId())
                 .nickname(nickname)
                 .profileImgUrl(user.getProfileImgUrl())
-                // TODO: user_review 테이블 구현 후 추가
-                // .rating(null)
-                // .tags(List.of())
                 .build();
     }
 
@@ -57,8 +68,5 @@ public class NowTravelPostResponseDto {
         private Long userId;
         private String nickname;
         private String profileImgUrl;
-        // TODO: user_review 테이블 구현 후 추가
-        // private Double rating;
-        // private List<String> tags;
     }
 }
